pyproject.toml
[tool.poetry]
name = "naviproject"
version = "0.1.0"
description = ""
authors = ["kuptsovsasha <kuptsovsasha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
drf-yasg = "^1.21.3"
drf-yasg-stubs = "^0.1.4"
permission = "^0.4.1"
django-filter = "^22.1"
django-rest-framework = "^0.1.0"
django-environ = "^0.9.0"
django-factory_boy = "^1.0.0"
mypy = "^0.971"
isort = "^5.10.1"
black = "^22.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
    (
        \.eggs|
        \.git|
        \.hg|
        \.mypy_cache|
        \.nox|
        \.tox|
        \.venv|
        \.svn|
        _build|
        buck-out|
        build|
        dist|
        settings\.py|
        migrations|
        snapshots|

    )/
'''
